requirements.txt
face-recognition==1.3.0
numpy>=1.24.0
opencv-python>=4.8.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
pillow>=10.0.0
pytest>=7.4.0  # for testing 